requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.0.40
openai>=1.0.0
plotly>=5.15.0
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pytest>=7.4.0
requests>=2.31.0
beautifulsoup4>=4.12.0
yfinance>=0.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
jupyter>=1.0.0
notebook>=6.5.0
python-multipart>=0.0.6
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
typing-extensions>=4.7.0

